pyproject.toml
[project]
name = "sem-denoising"
version = "1.0.0"
description = "SEM Image Denoising Application with PySide6"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scikit-image>=0.22.0",
    "scipy>=1.11.0",
    "Pillow>=10.0.0",
    "PyWavelets>=1.4.0",
]

[tool.uv]
dev-dependencies = []
